ref: move test files to e2e/integration
- cypress has the same directory structure
