+++ packages/schematics/e2e/integration/jest.spec.ts
@@ -1,9 +1,9 @@
 import {
   IntegrationTestBed,
   initializeIntegrationTestBed
-} from './integration-test-bed';
+} from '../integration-test-bed';
 
-import { linkSchematics, matchLines } from './utils';
+import { linkSchematics, matchLines } from '../utils';
 
 describe('Jest', () => {
   let testBed: IntegrationTestBed;
